+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/STS/Utxo.hs
@@ -392,7 +392,7 @@ utxoInductive = do
         filter
           ( \x ->
               let c = getField @"value" x
-               in (Val.coin c) < minUTxOValue
+               in (Val.coin c) < (Val.scaledMinDeposit c minUTxOValue)
           )
           outputs
   null outputsTooSmall ?! OutputTooSmallUTxO outputsTooSmall
